requirements.txt
openai==1.86.0
langchain==0.3.25
langchain-openai==0.3.22
langgraph==0.4.8
python-dotenv==1.1.0
faiss-cpu==1.13.2
langchain-community==0.3.25
pypdf==5.6.0